requirements.txt
flask==2.2.2
Werkzeug==2.2.2
flask-sqlalchemy==2.5.1
flask-session
flask-migrate==4.0.7
python-dotenv
requests
gunicorn
psycopg2==2.9.10